requirements.txt
# Core dependencies
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.20.0
pillow>=9.0.0
pyyaml>=6.0
matplotlib>=3.5.0
opencv-python>=4.5.0
tqdm>=4.60.0
scikit-learn>=1.0.0

# For visualization
vtk>=9.0.0